pyproject.toml
[tool.poetry]
name = "processing-image"
version = "0.1.0"
description = ""
authors = ["azamtoiri <azamtoiri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.31"
celery = {extras = ["redis"], version = "^5.4.0"}
fastapi = "^0.111.0"
pillow = "^10.3.0"
loguru = "^0.7.2"
types-aiobotocore = {extras = ["essential"], version = "^2.13.1"}
aiobotocore = "^2.13.1"
aiofiles = "^24.1.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
redis = "^5.0.7"
flower = "^2.0.1"
requests = "^2.32.3"
fastapi-cache2 = "^0.2.1"
retry = "^0.9.2"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
